Refactored humaninterface to humaninput.
