+++ src/org/usfirst/frc4904/robot/humaninput/drivers/Nathan.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.humaninterface.drivers;
+package org.usfirst.frc4904.robot.humaninput.drivers;
 
 
 import org.usfirst.frc4904.robot.DriverStationMap;
@@ -6,18 +6,18 @@ import org.usfirst.frc4904.robot.RobotMap;
 import org.usfirst.frc4904.standard.commands.Kill;
 import org.usfirst.frc4904.standard.commands.chassis.ChassisIdle;
 import org.usfirst.frc4904.standard.commands.chassis.ChassisShift;
-import org.usfirst.frc4904.standard.humaninterface.Driver;
+import org.usfirst.frc4904.standard.humaninput.Driver;
 import org.usfirst.frc4904.standard.subsystems.chassis.SolenoidShifters;
 
-public class JoystickControl extends Driver {
-	public JoystickControl() {
-		super("JoystickControl");
+public class Nathan extends Driver {
+	public Nathan() {
+		super("Nathan"); // supernathan!
 	}
 	
 	public void bindCommands() {
 		DriverStationMap.xbox.back.whenPressed(new Kill(new ChassisIdle(RobotMap.chassis)));
-		DriverStationMap.xbox.rb.whenPressed(new ChassisShift(RobotMap.chassis.getShifter(), SolenoidShifters.ShiftState.DOWN));
-		DriverStationMap.xbox.lb.whenPressed(new ChassisShift(RobotMap.chassis.getShifter(), SolenoidShifters.ShiftState.UP));
+		DriverStationMap.xbox.a.whenPressed(new ChassisShift(RobotMap.chassis.getShifter(), SolenoidShifters.ShiftState.DOWN));
+		DriverStationMap.xbox.b.whenPressed(new ChassisShift(RobotMap.chassis.getShifter(), SolenoidShifters.ShiftState.UP));
 	}
 	
 	public double getX() {
@@ -25,10 +25,10 @@ public class JoystickControl extends Driver {
 	}
 	
 	public double getY() {
-		return DriverStationMap.xbox.leftStick.getY();
+		return DriverStationMap.xbox.rt.getX() - DriverStationMap.xbox.lt.getX();
 	}
 	
 	public double getTurnSpeed() {
-		return DriverStationMap.xbox.rightStick.getX();
+		return DriverStationMap.xbox.leftStick.getX();
 	}
 }
