+++ src/org/usfirst/frc4904/robot/humaninput/drivers/PureStick.java
@@ -1,10 +1,10 @@
-package org.usfirst.frc4904.robot.humaninterface.drivers;
+package org.usfirst.frc4904.robot.humaninput.drivers;
 
 
 import org.usfirst.frc4904.robot.DriverStationMap;
 import org.usfirst.frc4904.robot.RobotMap;
 import org.usfirst.frc4904.standard.commands.chassis.ChassisShift;
-import org.usfirst.frc4904.standard.humaninterface.Driver;
+import org.usfirst.frc4904.standard.humaninput.Driver;
 import org.usfirst.frc4904.standard.subsystems.chassis.SolenoidShifters;
 
 public class PureStick extends Driver {
